langchain-react-demo/requirements.txt
fastapi
uvicorn
openai
langchain
langchain-community
chromadb
tqdm
pydantic
requests
tiktoken
python-dotenv